Add usability for io.load_cube to be used with a list of paths
This also tests the loaded cube(s) for the number of coordinates, in
case no name constraint is given. This will discard additional cubes
from a file that are not a likely dataset with 1+3 coordinates.

This commit further removes the time selection part: the full time
range is read in and used.