Add a simple 'at-list' module
This lets a user supply an input file with list of filenames, instead
of listing all filenames on the command line. This allows an easier
overview of the input files, and especially easier modification (for
example, by commenting out certain files).

The kcs.cmip.distribution module has been changed to accept at-lists.