Grab variable name from filename for data loading
For some datasets, Iris needs a constraint, or it will crash upon
loading (as other variables appear to get in the way).

Grabbing the variable name from the filename works for standard CMIP 5
or 6 data, but since it requires the variable (short) name to be at
the start of the filename, it may not always work. If no variable
short name is found, there will be no constraint; this may still work
for most datasets.