Add a configuration file
The configuration file should, once complete, contain all major and
minor settings, including defaults for command line options. Comments
should guide a user to adjusting the configuration.

The configuration file is in TOML format.

For ease, the file can be split into several files, for the main
tables (sections). These are included using the special "include" key
(this is not a TOML standard).