Add a default configuration
Rewrite the `config` subpackage to read a default configuration, from
a TOML string. The TOML string can also be printed to stdout by
calling the subpackage as a module: `python -m kcs.config`. Using the
configuration as a TOML string keeps both the comments and the format
(a nested dict printed to TOML format with `toml.dumps(config_dict)`
may re-order the sections in the configuration).